fix: remove SetWindowFocus() in Render() when scrolling to cursor
When the text editor is not rendered into the currently focused window, this provoke current windows to switch. This is a problem because "scroll to cursor" do not require any focus changes. I might be usefull if window is hidden, but if it is the case, the UI user should be able to do it manually.